+++ crates/fluvio-test/src/tests/producer_fail/mod.rs
@@ -1,6 +1,6 @@
 use std::any::Any;
 
-use fluvio::{RecordKey, TopicProducer, FluvioAdmin, FluvioError};
+use fluvio::{RecordKey, TopicProducer, TopicProducerConfigBuilder, FluvioAdmin, FluvioError};
 use fluvio_controlplane_metadata::partition::PartitionSpec;
 use clap::Parser;
 
@@ -47,7 +47,14 @@ pub async fn produce_batch(
     println!("Starting produce_batch test");
 
     let topic_name = test_case.environment.base_topic_name();
-    let producer: TopicProducer = test_driver.create_producer(&topic_name).await;
+    let config = TopicProducerConfigBuilder::default()
+        .linger(std::time::Duration::from_millis(10))
+        .build()
+        .expect("failed to build config");
+
+    let producer: TopicProducer = test_driver
+        .create_producer_with_config(&topic_name, config)
+        .await;
 
     println!("Created producer");
 
@@ -71,10 +78,11 @@ pub async fn produce_batch(
 
     println!("Got cluster manager");
 
+    let value = "a".repeat(5000);
     let result: Result<_, FluvioError> = (|| async move {
         let mut results = Vec::new();
-        for i in 0..1000 {
-            let result = producer.send(RecordKey::NULL, i.to_string()).await?;
+        for _ in 0..1000 {
+            let result = producer.send(RecordKey::NULL, value.clone()).await?;
             results.push(result);
         }
         println!("Send 1000");